AtmClient renamed into AtmClientSingleton
